+++ R/tm_squad_stats.R
@@ -12,6 +12,8 @@
 #' @export
 tm_squad_stats <- function(team_url) {
 
+  main_url <- "https://www.transfermarkt.com"
+
   # .pkg_message("Scraping squad player stats. Please acknowledge transfermarkt.com as the data source")
 
   each_squad_stats <- function(each_team_url) {
@@ -29,6 +31,7 @@ tm_squad_stats <- function(team_url) {
       team_data_table <- team_data_page %>% rvest::html_nodes("#yw1") %>% rvest::html_node("table") %>% rvest::html_nodes("tbody") %>% rvest::html_children()
 
       player_name <- team_data_table %>% rvest::html_nodes(".hauptlink") %>% rvest::html_nodes(".hide-for-small") %>% rvest::html_text()
+      player_url <- team_data_table %>% rvest::html_nodes(".hauptlink") %>% rvest::html_nodes(".hide-for-small a") %>% rvest::html_attr("href") %>% paste(main_url, .)
       player_pos <- team_data_table %>% rvest::html_nodes(".inline-table tr+ tr td") %>% rvest::html_text()
       player_age <- team_data_table %>% rvest::html_nodes(".posrela+ .zentriert") %>% rvest::html_text()
       nationality <- team_data_table %>% rvest::html_nodes(".flaggenrahmen:nth-child(1)") %>% rvest::html_attr("title")
@@ -41,9 +44,10 @@ tm_squad_stats <- function(team_url) {
       minutes_played <- team_data_table %>% rvest::html_nodes(".rechts") %>% rvest::html_text() %>%
         gsub("\\.", "", .) %>% gsub("'", "", .) %>% gsub("-", "0", .) %>% as.numeric()
 
-      team_data_df <- data.frame(player_name = as.character(player_name), player_pos = as.character(player_pos), player_age = as.numeric(player_age),
-                                 nationality = as.character(nationality), in_squad = as.numeric(in_squad), appearances = as.numeric(appearances),
-                                 goals = as.numeric(goals), minutes_played = as.numeric(minutes_played))
+      team_data_df <- data.frame(team_name = as.character(team_name), league = as.character(league), country = as.character(country),
+                                 player_name = as.character(player_name), player_url = as.character(player_url), player_pos = as.character(player_pos),
+                                 player_age = as.numeric(player_age), nationality = as.character(nationality), in_squad = as.numeric(in_squad),
+                                 appearances = as.numeric(appearances), goals = as.numeric(goals), minutes_played = as.numeric(minutes_played))
 
     } else {
       team_data_df <- data.frame()
@@ -56,9 +60,10 @@ tm_squad_stats <- function(team_url) {
   # create the progress bar with a progress function.
   pb <- progress::progress_bar$new(total = length(team_url))
 
-  final_output <- team_url %>%
-    purrr::map_df(each_squad_stats)
-
-  return(final_output)
+  f_possibly <- purrr::possibly(each_squad_stats, otherwise = data.frame(), quiet = FALSE)
+  purrr::map_dfr(
+    team_url,
+    f_possibly
+  )
 
 }